+++ quadratic-core/src/controller/user_actions/cells.rs
@@ -125,34 +125,47 @@ mod test {
             y: 6,
             sheet_id,
         };
-        let get_the_cell = |g: &GridController| {
+        let get_cell = |g: &GridController| {
             g.sheet(sheet_id)
                 .get_cell_value(sheet_pos.into())
                 .unwrap_or_default()
         };
         let mut cell_sheets_modified = HashSet::new();
         cell_sheets_modified.insert(CellSheetsModified::new(sheet_pos));
-        assert_eq!(get_the_cell(&g), CellValue::Blank);
-        g.set_cell_value(sheet_pos, String::from("a"), None);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("a")));
-        g.set_cell_value(sheet_pos, String::from("b"), None);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("b")));
+        assert_eq!(get_cell(&g), CellValue::Blank);
+
+        // test undo/redo of a single cell value and ensure that cell_sheets_modified is properly populated for the renderer
+        let summary = g.set_cell_value(sheet_pos, String::from("a"), None);
+        assert_eq!(summary.cell_sheets_modified, cell_sheets_modified);
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("a")));
+
+        let summary = g.set_cell_value(sheet_pos, String::from("b"), None);
+        assert_eq!(summary.cell_sheets_modified, cell_sheets_modified);
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("b")));
+
         assert_eq!(g.undo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("a")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("a")));
+
         assert_eq!(g.redo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("b")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("b")));
+
         assert_eq!(g.undo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("a")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("a")));
+
         assert_eq!(g.undo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Blank);
+        assert_eq!(get_cell(&g), CellValue::Blank);
+
         assert_eq!(g.undo(None).cell_sheets_modified, HashSet::default());
-        assert_eq!(get_the_cell(&g), CellValue::Blank);
+        assert_eq!(get_cell(&g), CellValue::Blank);
+
         assert_eq!(g.redo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("a")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("a")));
+
         assert_eq!(g.redo(None).cell_sheets_modified, cell_sheets_modified);
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("b")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("b")));
+
         assert_eq!(g.redo(None).cell_sheets_modified, HashSet::default());
-        assert_eq!(get_the_cell(&g), CellValue::Text(String::from("b")));
+        assert_eq!(get_cell(&g), CellValue::Text(String::from("b")));
     }
 
     #[test]
